requirements.txt
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0
requests==2.32.3
